telegramforwarder_backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
telethon==1.29.2
python-dotenv==1.0.0